+++ tests/invite_legacy_test.go
@@ -18,20 +18,24 @@ import (
 	refs "go.mindeco.de/ssb-refs"
 
 	"go.cryptoscope.co/ssb/client"
+	"go.cryptoscope.co/ssb/internal/testutils"
 	"go.cryptoscope.co/ssb/invite"
 )
 
-// TODO: enable plugin
-
 // first js creates an invite
 // go will try to use it
-func XTestLegacyInviteJSCreate(t *testing.T) {
+func TestLegacyInviteJSCreate(t *testing.T) {
+	if testutils.SkipOnCI(t) {
+		return
+	}
+
 	r := require.New(t)
 
 	os.Remove("legacy_invite.txt")
 
-	// ts := newRandomSession(t)
 	ts := newSession(t, nil, nil)
+	// TODO: configure ssbClient to use the right shs-cap
+	// ts := newRandomSession(t)
 
 	ts.startGoBot()
 	bob := ts.gobot
@@ -40,7 +44,7 @@ func XTestLegacyInviteJSCreate(t *testing.T) {
 	// manual multiserver address
 	addr := fmt.Sprintf("net:%s", netwrap.GetAddr(wrappedAddr, "tcp").String())
 	addr += "~shs:"
-	addr += base64.StdEncoding.EncodeToString(bob.KeyPair.Id.ID)
+	addr += base64.StdEncoding.EncodeToString(bob.KeyPair.Id.PubKey())
 	t.Log("addr:", addr)
 
 	bob.PublishLog.Append(map[string]interface{}{
@@ -117,8 +121,9 @@ func XTestLegacyInviteJSCreate(t *testing.T) {
 	<-ts.doneJS
 
 	// now follow alice
-	_, err = bob.PublishLog.Publish(refs.NewContactFollow(&tok.Peer))
+	_, err = bob.PublishLog.Publish(refs.NewContactFollow(tok.Peer))
 	r.NoError(err)
+	bob.Replicate(alice)
 
 	hasBobsFeed := `
 	sbot.on('rpc:connect', (rpc) => {
@@ -157,12 +162,13 @@ func XTestLegacyInviteJSCreate(t *testing.T) {
 	ts.wait()
 }
 
-func XTestLegacyInviteJSAccept(t *testing.T) {
+func TestLegacyInviteJSAccept(t *testing.T) {
 	r := require.New(t)
 
 	os.Remove("legacy_invite.txt")
 
-	// ts := newRandomSession(t) // TODO: configure ssbClient to use the right shs-cap
+	// TODO: configure ssbClient to use the right shs-cap
+	// ts := newRandomSession(t)
 	ts := newSession(t, nil, nil)
 
 	ts.startGoBot()
@@ -172,7 +178,7 @@ func XTestLegacyInviteJSAccept(t *testing.T) {
 	// manual multiserver address
 	addr := fmt.Sprintf("net:%s", netwrap.GetAddr(wrappedAddr, "tcp").String())
 	addr += "~shs:"
-	addr += base64.StdEncoding.EncodeToString(bob.KeyPair.Id.ID)
+	addr += base64.StdEncoding.EncodeToString(bob.KeyPair.Id.PubKey())
 	t.Log("addr:", addr)
 
 	bob.PublishLog.Append(map[string]interface{}{
